Various action improvements
This commit adds logic that checks whether the current run of the
workflow and the previous run have the same conclusion, and based on the
newly added `NOTIFY_ONLY_ON_CONCLUSION_CHANGE` setting it will notify or
not. This is useful to avoid notifications being spammed.

This commit also adds a `EXTRA_INFORMATION` setting which allows you to
add custom data in your notification and various code
improvements/comments.

All these changes are also reflected in the readme.